+++ src/server.ts
@@ -2,11 +2,11 @@ import { Device, Discovery } from './';
 
 const discovery = new Discovery();
 
-discovery.onHello((device: Device) => {
+discovery.on('hello', (device: Device) => {
     console.log(`${new Date().toLocaleTimeString()} - Hello from ${device.macAddress} on ${device.address}`);
 });
 
-discovery.onGoodbye((device: Device) => {
+discovery.on('goodbye', (device: Device) => {
     console.log(`${new Date().toLocaleTimeString()} - Goodbye from ${device.macAddress} on ${device.address}`);
 });